UI Elements created
Scripts for the player score, time and name were created and linked to the text handler.

The text handler's job is to update the HUD text.

The time also goes to 2 decimal places (initially I used math to do this, then I realized it messed with the speed, so I simply reduced the size of the text box. This won't work as well if the text is increased by another decimal, but it should be fine for this.