Built Level 1*, Worked on camera.. Alas ran into problem
I rebuilt the second half of the first level and then went to make the camera follow the player.
But when the player changes direction the camera follows suit. I looked through the camera settings and so far haven't found a way to stop this.
I could easily write a script to make the camera follow the character on the axis I want and avoid cinemachine all together. But cinemachine is good so I don't want to do that.