Added Stage Selection Draft + Build Update
When the player wins a level, the stage level will increment.

When the player loads the stage selection screen, a script runs through all the stage buttons. If the button (which contains a Level ID) returns a ID higher than the current highest level, it is deactivated.

Talking about UI, I use a UI group handler which automatically moves the selected UI into a grid formation.

As I was just about to build and commit, I happened to find a bug.... which meant I continued for another hour or so trying to fix it.. ending up not being able to :T